clean up workflow js files
