+++ workflow-scripts/constants.workflow.mjs
@@ -1,0 +1,1 @@
+export const CONFIG_PATH = './workflow.config.json';
